[fenix] Closes https://github.com/mozilla-mobile/fenix/issues/20679: revert allowaccessmodification proguard disabling.
We disabled the allowaccessmodification proguard option because it broke
functionality or crashed the app (I can't rememeber). As far as we know,
the R8 bug was fixed in the R8 bundled with the Android Gradle Plugin
v4.1. We're now on AGP v7.0.0-rc1 so we should be able to revert this
now.

This commit reverts the following commits:

Revert "Proguard/r8: Do not allow access modification."
This reverts commit d2ec7c648856664c27b31831959fd2e83a580968.

Revert "Dump `proguard-android-optimize.txt` into local configuration for later modification"
This reverts commit c543ae338e0bce4a6e2395f3e72742d9c0d65042.